M2-5555: correction of code review comments
