+++ src/modules/Dashboard/features/RespondentData/RespondentDataSummary/Report/Subscales/Subscales.utils.ts
@@ -1,8 +1,11 @@
 import { ActivitySettingsSubscale } from 'shared/state';
 import { ActivityItemAnswer, ElementType } from 'shared/types';
+import {
+  ActivityCompletion,
+  FormattedResponse,
+} from 'modules/Dashboard/features/RespondentData/RespondentData.types';
 
 import { compareActivityItem, formatActivityItemAnswers } from '../Report.utils';
-import { ActivityCompletion, FormattedResponse } from '../../../RespondentData.types';
 import { ActivityCompletionToRender, GroupedSubscales, SubscaleToRender } from './Subscales.types';
 
 export const getSubscalesToRender = (
@@ -31,10 +34,10 @@ export const getSubscalesToRender = (
       }
     } else if (activityItems[item.name]) {
       const formattedItem = formatActivityItemAnswers(activityItems[item.name], endDatetime);
-      if (!result?.[data.name]?.items) {
-        result[data.name] = { items: [formattedItem] };
-      } else {
+      if (result?.[data.name]?.items) {
         result[data.name].items?.push(formattedItem);
+      } else {
+        result[data.name] = { items: [formattedItem] };
       }
     }
   }
@@ -62,13 +65,13 @@ export const getAllSubscalesToRender = (
           item.endDatetime,
         );
 
-        if (!allSubscalesToRender?.[subscale.name]?.items) {
+        if (allSubscalesToRender?.[subscale.name]?.items) {
+          allSubscalesToRender[subscale.name].items?.push(formattedItem);
+        } else {
           allSubscalesToRender[subscale.name] = {
             ...allSubscalesToRender[subscale.name],
             items: [formattedItem],
           };
-        } else {
-          allSubscalesToRender[subscale.name].items?.push(formattedItem);
         }
       } else {
         allSubscalesToRender[subscale.name] = {
@@ -96,11 +99,13 @@ export const getAllSubscalesToRender = (
       }
 
       const { activityItem, answers } = compareActivityItem(
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         allSubscalesToRender[subscale.name].items![itemIndex],
         activityItems[subscaleItem.name],
         item.endDatetime,
       );
 
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       allSubscalesToRender[subscale.name].items![itemIndex] = {
         activityItem,
         answers,
